do not display admin menu if it has no children
